backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
redis==5.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.2
openai==1.45.0
anthropicai==0.34.2
websockets==13.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
aiofiles==24.1.0
